o Fixed an integer overflow which prevented a target specification   of "*.*.*.*" from working.  Support for the CIDR /0 is now also   available. [Kris]
A sendto() or connect() to IPv4 addresses with a first octet of 0
(excluding 0.0.0.0 on some OS's) returns EINVAL.  These addresses
are reserved for "self-identification" and should never go over the
wire (and at least on Linux a connect() to 0.0.0.0 gets redirected
to 127.0.0.1).  I'm talking with Fyodor on whether or not to skip
*all* of these addresses.

+++ TargetGroup.h
@@ -160,7 +160,7 @@ class TargetGroup {
 
 /* Number of IPs left in this structure -- set to 0 if 
 		  the fields are not valid */
-  unsigned long ipsleft; 
+  unsigned long long ipsleft; 
 
   // is the current target expression a named host
   int namedhost;
